Marker is now found in PTAM world CS.
The AR toolkit is now only used to find the positions
in the image of the marker corners. The corners is then
projected onto the PTAM ground plane.

The points can later be used to calculate the scale
and orientation of the PTAM world. This is not done
right done.