Refactored the MapMaker and removed support for multimaps
Removed the feature of handling multiple maps because its not
needed at all and adds a lot of "noise" everywhere in the code
base since every class needs to be able to handle the switching.

The MapMaker is also refactored completly and have a much smaller
and easier to use interface.